Init the dashboard angular project as well
Changed from json to msgpack and added bunjs, removed go

Added msgpack files in include and removed json. This was done to
improve proformance and also to move querying from server to the db
engine. It is easier to convert msgpack to navtive types.. I think, I
mean I just read it somewhere so it must me right... right? Well it
hopefully should be easier to implement querying now with this change.
One thing thats is better with this is storage it should reduce file
size with alot which I hope also helps when doing large greps with alot
of files. Also a big BONUS is that data types should now also be saved,
WOOOOO!!

Anyway also cleaned up so that the files now has its own folders in src
making it more clean than just one file :D

Also added CMake to compile cuz i wanna :D

=======================================================================

On the server side I changed out GO for bunJS and I know its not as good
and fast as GO but it all came down to that I am way more comfy using
typescript and js comparred to GO. I also choose bunjs becuase it has a
good c api connection comparred to deno and nodejs and I think that is
because it is written in zig(Intressting language btw). Anyway I removed
the GO files and added .ts files instead and started testing the C api
package to learn it abit before I full on build out the express server.

========================================================================

Im gonna builed out the backend of things before starting to even think
about building the dashboard in Angular.

========================================================================

Signed Of By :
GrimFilerino